chore: refactor tracking
