--- src/const/abi/jumperApiUrls.ts
@@ -1,3 +1,0 @@
-// jumper api endpoints
-export const JUMPER_ANALYTICS_EVENT = '/users/events';
-export const JUMPER_ANALYTICS_TRANSACTION = '/wallets/transactions';
