+++ src/hooks/userTracking/useUserTracking.ts
@@ -1,7 +1,10 @@
 'use client';
 import { getSiteUrl } from '@/const/urls';
-import type { JumperEventData } from '@/hooks/useJumperTracking';
-import { useJumperTracking } from '@/hooks/useJumperTracking';
+import type { JumperEventData } from '@/utils/tracking/jumperTracking';
+import {
+  trackJumperEvent,
+  trackJumperTransaction,
+} from '@/utils/tracking/jumperTracking';
 import { useSession } from '@/hooks/useSession';
 import type {
   TrackEventProps,
@@ -12,7 +15,7 @@ import { EventTrackingTool } from '@/types/userTracking';
 import { useAccount } from '@lifi/wallet-management';
 import type { Theme } from '@mui/material';
 import { useMediaQuery } from '@mui/material';
-import { useCallback } from 'react';
+import { useCallback, useMemo } from 'react';
 import { TrackingEventParameter } from 'src/const/trackingKeys';
 import { useAbTestsStore } from 'src/stores/abTests';
 import { useFeatureFlags } from 'src/hooks/useFeatureFlags';
@@ -73,6 +76,49 @@ const addressableEvent = ({
     );
 };
 
+function buildTransactionPayload(data: TrackTransactionDataProps) {
+  return {
+    errorCode: data[TrackingEventParameter.ErrorCode],
+    errorCodeKey: data[TrackingEventParameter.ErrorCodeKey],
+    errorMessage: data[TrackingEventParameter.ErrorMessage],
+    exchange: data[TrackingEventParameter.Exchange],
+    feeCost: data[TrackingEventParameter.FeeCost],
+    feeCostFormatted: data[TrackingEventParameter.FeeCostFormatted],
+    feeCostUSD: data[TrackingEventParameter.FeeCostUSD],
+    fromAmount: Number(data[TrackingEventParameter.FromAmount]),
+    fromAmountUSD: Number(data[TrackingEventParameter.FromAmountUSD]),
+    fromChainId: data[TrackingEventParameter.FromChainId],
+    fromToken: data[TrackingEventParameter.FromToken],
+    gasCost: data[TrackingEventParameter.GasCost],
+    gasCostFormatted: data[TrackingEventParameter.GasCostFormatted],
+    gasCostUSD: Number(data[TrackingEventParameter.GasCostUSD]),
+    integrator: data[TrackingEventParameter.Integrator],
+    isFinal: data[TrackingEventParameter.IsFinal],
+    lastStepAction: data[TrackingEventParameter.LastStepAction],
+    message: data[TrackingEventParameter.Message],
+    nbOfSteps: data[TrackingEventParameter.NbOfSteps],
+    routeId: data[TrackingEventParameter.RouteId],
+    slippage: data[TrackingEventParameter.Slippage],
+    maxSlippage: data[TrackingEventParameter.MaxSlippage],
+    status: data[TrackingEventParameter.Status],
+    stepIds: data[TrackingEventParameter.StepIds],
+    steps: data[TrackingEventParameter.Steps],
+    tags: data[TrackingEventParameter.Tags],
+    time: data[TrackingEventParameter.Time],
+    toAmount: Number(data[TrackingEventParameter.ToAmount]),
+    toAmountFormatted: data[TrackingEventParameter.ToAmountFormatted],
+    toAmountMin: Number(data[TrackingEventParameter.ToAmountMin]),
+    toAmountUSD: Number(data[TrackingEventParameter.ToAmountUSD]),
+    toChainId: data[TrackingEventParameter.ToChainId],
+    toToken: data[TrackingEventParameter.ToToken],
+    transactionHash: data[TrackingEventParameter.TransactionHash],
+    transactionId: data[TrackingEventParameter.TransactionId],
+    transactionLink: data[TrackingEventParameter.TransactionLink],
+    transactionStatus: data[TrackingEventParameter.TransactionStatus],
+    type: data[TrackingEventParameter.Type],
+  };
+}
+
 export interface UserTracking {
   trackEvent: (event: TrackEventProps) => Promise<void>;
   trackTransaction: (transaction: TrackTransactionProps) => Promise<void>;
@@ -87,13 +133,18 @@ export function useUserTracking(): UserTracking {
   const sessionId = useSession();
   const fp = useFingerprint();
   const { activeAbTests } = useAbTestsStore();
-
-  const {
-    trackEvent: jumperTrackEvent,
-    trackTransaction: jumperTrackTransaction,
-  } = useJumperTracking();
-
   const { data: featureFlags } = useFeatureFlags();
+
+  const sessionContext = useMemo(
+    () => ({
+      browserFingerprint: fp || 'unknown',
+      walletAddress: account.address || 'not_connected',
+      walletProvider: account.connector?.name,
+      sessionId: sessionId || 'unknown',
+      abtests: activeAbTests,
+    }),
+    [fp, account.address, account.connector?.name, sessionId, activeAbTests],
+  );
 
   const getFeatureFlagVariants = useCallback(
     (eventName: string): Record<string, string | boolean> => {
@@ -126,52 +177,32 @@ export function useUserTracking(): UserTracking {
         googleEvent({ action, category, data });
       }
       if (enableAddressable) {
-        const dataArray = [];
-        if (label) {
-          dataArray.push({ name: 'label', value: label });
-        }
-
         addressableEvent({ action, label, data: data || {}, isConversion });
       }
       if (!disableTrackingTool?.includes(EventTrackingTool.JumperTracking)) {
         try {
           const flagVariants = getFeatureFlagVariants(action);
-          const eventData = {
-            category,
-            value: typeof value === 'number' ? value : 0,
+          await trackJumperEvent({
+            ...sessionContext,
             action,
+            category,
             label,
             data: data || {},
+            value: typeof value === 'number' ? value : 0,
             isConnected: account.isConnected || false,
-            walletAddress: account.address || 'not_connected',
-            walletProvider: account.connector?.name,
-            browserFingerprint: fp || 'unknown',
             isMobile: !isDesktop,
-            sessionId: sessionId || 'unknown',
             referrer: document?.referrer,
             url: window?.location?.href || getSiteUrl(),
-            abtests: activeAbTests,
             ...(Object.keys(flagVariants).length > 0 && {
               abTestVariants: flagVariants,
             }),
-          };
-          await jumperTrackEvent(eventData);
+          });
         } catch (error) {
           console.error('Error in tracking event:', error);
         }
       }
     },
-    [
-      account.address,
-      account.connector?.name,
-      account.isConnected,
-      fp,
-      isDesktop,
-      jumperTrackEvent,
-      sessionId,
-      activeAbTests,
-      getFeatureFlagVariants,
-    ],
+    [sessionContext, account.isConnected, isDesktop, getFeatureFlagVariants],
   );
 
   const trackTransaction = useCallback(
@@ -188,59 +219,15 @@ export function useUserTracking(): UserTracking {
       if (!disableTrackingTool?.includes(EventTrackingTool.JumperTracking)) {
         const dataAction = data[TrackingEventParameter.Action] ?? action ?? '';
         const flagVariants = getFeatureFlagVariants(dataAction);
-        const transactionData = {
-          url: window?.location?.href || getSiteUrl(),
-          browserFingerprint: fp || 'unknown',
-          walletAddress: account.address || 'not_connected',
-          walletProvider: account.connector?.name,
-          referrer: document?.referrer,
-          abtests: activeAbTests,
-          // data from handleRouteTrackingData:
+        await trackJumperTransaction({
+          ...sessionContext,
+          ...buildTransactionPayload(data),
           action: dataAction,
-          errorCode: data[TrackingEventParameter.ErrorCode],
-          errorCodeKey: data[TrackingEventParameter.ErrorCodeKey],
-          errorMessage: data[TrackingEventParameter.ErrorMessage],
-          exchange: data[TrackingEventParameter.Exchange],
-          feeCost: data[TrackingEventParameter.FeeCost],
-          feeCostFormatted: data[TrackingEventParameter.FeeCostFormatted],
-          feeCostUSD: data[TrackingEventParameter.FeeCostUSD],
-          fromAmount: Number(data[TrackingEventParameter.FromAmount]),
-          fromAmountUSD: Number(data[TrackingEventParameter.FromAmountUSD]),
-          fromChainId: data[TrackingEventParameter.FromChainId],
-          fromToken: data[TrackingEventParameter.FromToken],
-          gasCost: data[TrackingEventParameter.GasCost],
-          gasCostFormatted: data[TrackingEventParameter.GasCostFormatted],
-          gasCostUSD: Number(data[TrackingEventParameter.GasCostUSD]),
-          integrator: data[TrackingEventParameter.Integrator],
-          isFinal: data[TrackingEventParameter.IsFinal],
-          lastStepAction: data[TrackingEventParameter.LastStepAction],
-          message: data[TrackingEventParameter.Message],
-          nbOfSteps: data[TrackingEventParameter.NbOfSteps],
-          routeId: data[TrackingEventParameter.RouteId],
-          sessionId: sessionId || '',
-          slippage: data[TrackingEventParameter.Slippage],
-          maxSlippage: data[TrackingEventParameter.MaxSlippage],
-          status: data[TrackingEventParameter.Status],
-          stepIds: data[TrackingEventParameter.StepIds],
-          steps: data[TrackingEventParameter.Steps],
-          tags: data[TrackingEventParameter.Tags],
-          time: data[TrackingEventParameter.Time],
-          toAmount: Number(data[TrackingEventParameter.ToAmount]),
-          toAmountFormatted: data[TrackingEventParameter.ToAmountFormatted],
-          toAmountMin: Number(data[TrackingEventParameter.ToAmountMin]),
-          toAmountUSD: Number(data[TrackingEventParameter.ToAmountUSD]),
-          toChainId: data[TrackingEventParameter.ToChainId],
-          toToken: data[TrackingEventParameter.ToToken],
-          transactionHash: data[TrackingEventParameter.TransactionHash],
-          transactionId: data[TrackingEventParameter.TransactionId],
-          transactionLink: data[TrackingEventParameter.TransactionLink],
-          transactionStatus: data[TrackingEventParameter.TransactionStatus],
-          type: data[TrackingEventParameter.Type],
+          url: window?.location?.href || getSiteUrl(),
+          referrer: document?.referrer,
           abTestVariants: { ...data.abTestVariants, ...flagVariants },
-        };
-        await jumperTrackTransaction(transactionData);
+        });
       }
-
       if (enableAddressable) {
         addressableEvent({
           action,
@@ -250,15 +237,7 @@ export function useUserTracking(): UserTracking {
         });
       }
     },
-    [
-      account.address,
-      account.connector?.name,
-      fp,
-      jumperTrackTransaction,
-      sessionId,
-      activeAbTests,
-      getFeatureFlagVariants,
-    ],
+    [sessionContext, getFeatureFlagVariants],
   );
 
   return {
